fix(processFeedGains): outputs should be float16

+++ lib/stages/processFeedGains.hpp
@@ -8,6 +8,7 @@
 #define PROCESS_FEED_GAINS_HPP
 
 #include "Config.hpp"          // for Config
+#include "DataType.hpp"        // for float16_t
 #include "Stage.hpp"           // for Stage
 #include "buffer.hpp"          // for Buffer
 #include "bufferContainer.hpp" // for bufferContainer
@@ -72,10 +73,10 @@ private:
 
     /// Implement logic to upchannelize a single fine frequency, given a coarse frequency.
     /// Default will just duplicate the coarse frequency gain `N` times.
-    virtual void copy_upchannelize_f(const float* src_f, float* dst_f, size_t fid);
+    virtual void copy_upchannelize_f(const float* src_f, float16_t* dst_f, size_t fid);
 
     /// Create a frame desc for the output buffer. Default creates a frame desc
-    /// containing [beam, freq, subfreq, element, ReIm] axes.
+    /// containing [beam, freq, element, ReIm] axes.
     virtual void set_frame_desc(Buffer* buf);
 
     std::vector<Buffer*> gain_buffers;
@@ -100,7 +101,7 @@ private:
     float scaling_factor;
 
     /// Fixed buffers used to hold gains separately from the kotekan buffers
-    std::vector<float> gain_store_buf;
+    std::vector<float16_t> gain_store_buf;
     std::vector<uint8_t> mask_store_buf;
 
     /// Store gain upchannelization factors
